Move whole form element into ProfileForm component
